Prefer maps over protocols, in some instances
Prior to this comment there were a few protocols that only existed for
the purpose of data access from the resource. This became reminiscent of
getter methods in Java/C#, and not the original intention behind
protocols.

This commit introduces a simplified multi-arity protocol
`ResourceProperties` with a single, multi-arity function called
`resource-properties`. The first arity provides the initial 'static'
resource properties, the second arity provides the resource properties
that are sensitive to the request. There may well be an overlap between
properties that are determined ahead-of-time and those determined upon
request. By returning maps in each case, the developer can decide which
property requires the request context and which do not, rather than
prescribing this through the original protocol function signatures.

A favorable outcome of this work has been the removal of the
ResourceFetch protocol, which was not well documented and had no tests
or examples. The reason is could be removed is that the developer can
now fetch as necessary in their implementation of
the (resource-properties resource ctx) function.

The original protocols have been removed (but a still commented out to
salvage the documentation, much of which is still relevant and needs to
be transferred to the new scheme).