Require yada.bidi in yada.yada
This ensures bidi protocol extension of yada.core/handler, which avoids
the need for clients to require yada.bidi. This is the intention behind
requiring yada.yada- to make things simple for users.

It should remain possible for other routing libraries to be integrated
with yada without conflicting with bidi. But users of yada.yada must be
prepared to accept recommended defaults, otherwise they should use
yada.core.