Add :authorization to access control with coercion
The current policy of allowing extra entries to appear in the
:access-control entry does not work for authorization, since
authorization happens at a per-realm basis. When the SingleRealm
coercion is applied to abbreviated forms, yada cannot know which sibling
entries in the abbreviated form should be copied into the
realm (e.g. :scheme), and which should be left in the :access-control
entry (e.g. :allow-origin).

This conundrum is solved by introducing an :authorization entry (on a
per-realm basis), which is known by the coercers but which can contain
any content.