Align with bidi 2.0.0 and expunge co-dependencies
Co-dependencies have finally been removed, thanks to the new uri-for
entry in bidi 2's match-context.

Although injection via the Ring request is not a general injection
solution, it does work well for yada due to the fact that there is
always a Ring request. The new approach still avoids hijacking the Ring
request while avoiding the complexity associated with co-dependencies.

This work prepares the ground for a proper write-up of the phonebook
example.