Add support for swagger namespaced keys to swagger output
This commit adds support for resource keys with the swagger namespace
being included in the swagger output with the namespace removed. This
means that yada now supports including swagger specific information
like tags in theresource without changing the resource Schema. This is
also future proof in that new swagger properties are supported without
having to change yada.

Fixes: #66