Add path parameters from bidi to swagger output
When the bidi route had path parameters but those parameters where not
defined in the resource the swagger output would have the path
parameter shown in the url but would be missing the definition for the
parameter. Similarly when the bidi route had qualifiers for the path
parameters you couldn't even generate the swagger output. This commit
fixes these issues so the swagger ouput now reflects what yada actually
does when handling a request.

When no path parameters are defined for a resource yada will simply
forward the path parameters from bidi directly with no check or
coercion. Meaning that bidi does all the type checking and coercion and
this commit will include the bidi path parameters with types in the
swagger output.

But when a resource does have path parameters defined these must cover
all path parameters and so overwrites whatever bidi has specified for
the parameters.

Fixes #41