Support wildcard in responses schema and disallow duplicates
While wildcard was allowed in get* the schema for responses didn't
allow it and so you couldn't use it.

There were also some corner cases in responses where the same status
could be in the key of several responses making it ambigous which
response was the correct one. So I have added a check to the schema
that requires that the keys are disjoint and so unambigous.