Allow method responses map to have integer and set keys
The responses map for the resource allows both integers, sets of
integers and * as keys while the responses documentation for the method
only allowed integer keys. This commit changes the method responses
documentation to allow the same keys as the resource responses map as
long as the map keys are disjoint.