Add responses map to swagger spec
The yada resource definition allowed a responses key on the resource
and a responses key for documentation on the method but none of those
where included in the swagger spec.

The spec now includes a merged version of the two responses maps in
the same way as is done for parameters. And since the swagger spec has
the concept of a default response and yada allowes * as a wildcard key
these are now mapped so that the wildcard key is documented as the
default response.