Add support for responses to swagger (#106)
* Add functions for working with responses map

The responses map in the resource can have integers, sets of integers
and * as keys and this commit adds methods that help with modifying
the map while kepping it disjoint.

* Allow method responses map to have integer and set keys

The responses map for the resource allows both integers, sets of
integers and * as keys while the responses documentation for the method
only allowed integer keys. This commit changes the method responses
documentation to allow the same keys as the resource responses map as
long as the map keys are disjoint.

* Add responses map to swagger spec

The yada resource definition allowed a responses key on the resource
and a responses key for documentation on the method but none of those
where included in the swagger spec.

The spec now includes a merged version of the two responses maps in
the same way as is done for parameters. And since the swagger spec has
the concept of a default response and yada allowes * as a wildcard key
these are now mapped so that the wildcard key is documented as the
default response.