Allow sub-resources to be less constrained
A sub-resource is really a dynamic resource determined at request time,
and perhaps should be named as such. It _can_ be used in conjunction
with path-info to produce a 'sub' resource (a resource 'below' another)
but really that's just one use-case of a dynamic resource.

This commit removes the constraint that sub-resources only be used with
path-info.