Assoc top level request headers into pieces
This change allows parts of multipart requests to access the headers for the surrounding request. This can be very useful for access user access tokens present in the request headers or any other contextual information pertaining to the request
as a whole that could influence the process of a part.