Move interceptors from resource to handler
The default interceptor chain (and error interceptor chain) is no longer
applied to the resource (although resources that specify interceptor
chains are still supported).

This helps raise the default interceptor chain up a level (to the
handler) to free yada.resource from dependencies (json, swagger,
json-html, etc.) that we are trying to untangle from the core.