Move interceptor chain to yada.yada
This is to allow re-use of yada.handler while delegating the definition
of the default interceptor chains to the potemkin wrapper namespace.

This allows a choice of which 'variant' of yada to import, either the
yada.yada variant which brings in all dependencies or the new 'lean'
variant which only brings in a minimal set of dependencies.

Also, renamed minimal to lean