Check referer header for same-origin? if origin not provided
Firefox and old IE versions don't provide an origin header when sending
POSTs. If a POST request from the same origin returns a URI, Yada
returns a 201 to these browsers, instead of a 303.

With this commit, Yada now checks the referer header to determine the
request's origin if the origin header is not provided.

Also remove lein alias for Travis CI as they no longer require it.