Revert recent changes to authentication/authorization
This is to abort the work that was done for both #249 and #245 in
favor of keeping binary compatibility with release 1.2.15. Release
1.2.16 is therefore marked as unsupported. Release 1.2.17 will
introduce a new authentication/authorization schema (by
accretion). Users that want to make use of the new design and
functionality will need to update their yada resources.

This seems the best course of action rather than to risk the situation
of old yada users upgrading and finding compatibility problems. The
new authentication will be easier for developers and with improved
conformance with the HTTP standards.

The security section in the manual is now wrong and will be updated
prior to the next release.

Fixes #251, but unfixes #249, #245