Implement new design for authentication and authorization
Documentation (security.adoc) has been updated to account for
this. The new design is compatible with the previous design since it
does not use the `:access-control` entry. This is therefore preseved
so that yada resources of the old design will still work in future
versions of yada. The old design in now considered frozen and new
features will only be added to the new design.

An important distinction is that the new design attempts to be
comformant with RFC 7235. This means that 'web' logins (using forms
and session cookies) are now not supported in the authentication
framework. Rather, they will be supported via support for RFC
6265 (HTTP State Management Mechanism).

Importantly, this work paves the way for supporting the Bearer
authentication scheme of RFC 6750 which will provide better support
for users wishing to create restricted-use APIs.