Add support for RFC 2616: HTTP State Management Mechanism
This is backward compatible with previous cookies support but adds the
declaration of cookies in yada resources. This makes it more
convenient to set cookies, which can be achieved by `yada/new-cookie`.

This commit also introduces the ability to return the yada context as
a return value to a response function. This is in order to make it
more straight-forward to perform cookie setting, and in future,
redirects and other HTTP behavior.