Rename :credentials for :authentication
The reasoning here is that credentials are often referred to as claims
made by the request, which after verification, become the
authenticated credentials. Using the term credentials for both pre and
post verified credentials is ambiguous. It is better to use the
`:authentication` key as it has symmetry with the `:authorization`
key, while preserving the distinction between authentication (the
verified identity of a user) and authorization (what the user is
allowed to do to the resource in context).