Preserve cookies after authentication failure
If cookies are unset by a cookie consumer (perhaps because they are
not valid), and authorization fails, the "set-cookie" header wasn't
being preserved in the error response. This means that there was no
easy way to unset a bad cookie.

This commit checks for cookies in the error handling code, and
preserves them if an exception was thrown.

It also adds a useful submap? function for testing part of the
response that we care about.