Limit logging of 501 Not Implemented errors
For public facing websites it's not uncommon to have exploitation suites run
random strings as methods in small batches of requests. This commit channels
501s away from errors logs into info logs as we may still want to know that
unknown methods are being used, typically though it's a part of the normal
course of operation and don't want it to fill logging with high priority
messages and potentially fall through into alerting for some users.